Agrgeagndo documentación al proyecto

+++ interfaz-grafica/IObjetoRemoto.java
@@ -6,5 +6,5 @@ public interface IObjetoRemoto extends Remote {
   // Firmas de los métodos de Cajero Automatico.
   Date lastMove(Date dateI) throws RemoteException;
   String Retirar(float cantidad) throws RemoteException;
-  void Depositar(int cantidad) throws RemoteException;
+  void Depositar(float cantidad) throws RemoteException;
 }
